chrome/test/data/xr/e2e_test_files/html/test_webxr_hand_tracking.html

<!doctype html>
<!--
A collection of helper functions and listeners to confirm the state of input
sources for the same object tests.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script src="../resources/webxr_boilerplate.js"></script>
    <script>
      function setupRequestHandTracking() {
        immersiveSessionInit = { requiredFeatures: ['hand-tracking'] }
      }

      let inputChangeEvents = 0;
      function onInputSourcesChange(event) {
        inputChangeEvents++;
      }

      onSessionStartedCallback = function(session) {
        if (session.mode.startsWith("immersive")) {
          session.addEventListener('inputsourceschange', onInputSourcesChange, false);
        }
      }

      function getCurrentInputSources() {
        let currentSession = sessionInfos[sessionTypes.IMMERSIVE].currentSession;
        return Array.from(currentSession.inputSources.values());
      }

      function assertHandTrackingFeatureState(expected_enabled) {
        let currentSession = sessionInfos[sessionTypes.IMMERSIVE].currentSession;
        let actual = currentSession.enabledFeatures.indexOf('hand-tracking') != -1;
        assert_equals(actual, expected_enabled);
      }

      function assertHandsNotPresent() {
        getCurrentInputSources().forEach((inputSource) => {
          assert_equals(inputSource.hand, null);
        });
      }

      function assertHandsPresent() {
        getCurrentInputSources().forEach((inputSource) => {
          assert_not_equals(inputSource.hand, null);
        });
      }

    </script>
  </body>
</html>
